Move selectBlock() to reduce compile time dependencies, re-factor. Optimise a couple of hotspots the profiler pointed out.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8445 3231ed4b-2112-0410-84e0-8c3944a40730

+++ HashLib/ED2K.h
@@ -1,7 +1,7 @@
 //
 // ED2K.h
 //
-// Copyright (c) Shareaza Development Team, 2002-2008.
+// Copyright (c) Shareaza Development Team, 2002-2010.
 // This file is part of SHAREAZA (shareaza.sourceforge.net)
 //
 // Shareaza is free software; you can redistribute it
@@ -66,4 +66,4 @@ private:
 	bool	m_bNullBlock;
 };
 
-const size_t ED2K_PART_SIZE	= 9500 * 1024u;
+const DWORD ED2K_PART_SIZE	= 9500 * 1024u;
